+++ tests_integration/nativeapp/test_deploy.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 import os
-import uuid
+from shlex import split
 
 from snowflake.cli.api.project.util import TEST_RESOURCE_SUFFIX_VAR
 from tests.nativeapp.utils import touch
@@ -44,8 +44,16 @@ def sanitize_deploy_output(default_username, resource_suffix):
 
 # Tests a simple flow of executing "snow app deploy", verifying that an application package was created, and an application was not
 @pytest.mark.integration
-@pytest.mark.parametrize("test_project", ["napp_init_v1", "napp_init_v2"])
+@pytest.mark.parametrize(
+    "command,test_project",
+    [
+        ["app deploy", "napp_init_v1"],
+        ["app deploy", "napp_init_v2"],
+        ["ws deploy --entity-id=pkg", "napp_init_v2"],
+    ],
+)
 def test_nativeapp_deploy(
+    command,
     test_project,
     nativeapp_project_directory,
     runner,
@@ -58,8 +66,9 @@ def test_nativeapp_deploy(
 ):
     project_name = "myapp"
     with nativeapp_project_directory(test_project):
-        result = runner.invoke_with_connection(["app", "deploy"])
+        result = runner.invoke_with_connection(split(command))
         assert result.exit_code == 0
+        assert "Validating Snowflake Native App setup script." in result.output
         assert sanitize_deploy_output(result.output) == snapshot
 
         # package exist
@@ -92,28 +101,74 @@ def test_nativeapp_deploy(
         )
 
         # re-deploying should be a no-op; make sure we don't issue any PUT commands
-        result = runner.invoke_with_connection_json(["app", "deploy", "--debug"])
+        result = runner.invoke_with_connection([*split(command), "--debug"])
         assert result.exit_code == 0
         assert "Successfully uploaded chunk 0 of file" not in result.output
 
 
 @pytest.mark.integration
 @pytest.mark.parametrize(
-    "command,contains,not_contains",
+    "command,contains,not_contains,test_project",
     [
         # deploy --prune removes remote-only files
         [
             "app deploy --prune --no-validate",
             ["stage/manifest.yml"],
             ["stage/README.md"],
+            "napp_init_v1",
+        ],
+        [
+            "app deploy --prune --no-validate",
+            ["stage/manifest.yml"],
+            ["stage/README.md"],
+            "napp_init_v2",
+        ],
+        [
+            "ws deploy --entity-id=pkg --prune --no-validate",
+            ["stage/manifest.yml"],
+            ["stage/README.md"],
+            "napp_init_v2",
         ],
         # deploy removes remote-only files (--prune is the default value)
-        ["app deploy --no-validate", ["stage/manifest.yml"], ["stage/README.md"]],
+        [
+            "app deploy --no-validate",
+            ["stage/manifest.yml"],
+            ["stage/README.md"],
+            "napp_init_v1",
+        ],
+        [
+            "app deploy --no-validate",
+            ["stage/manifest.yml"],
+            ["stage/README.md"],
+            "napp_init_v2",
+        ],
+        [
+            "ws deploy --entity-id=pkg --no-validate",
+            ["stage/manifest.yml"],
+            ["stage/README.md"],
+            "napp_init_v2",
+        ],
         # deploy --no-prune does not delete remote-only files
-        ["app deploy --no-prune", ["stage/README.md"], []],
+        [
+            "app deploy --no-prune",
+            ["stage/README.md"],
+            [],
+            "napp_init_v1",
+        ],
+        [
+            "app deploy --no-prune",
+            ["stage/README.md"],
+            [],
+            "napp_init_v2",
+        ],
+        [
+            "ws deploy --entity-id=pkg --no-prune",
+            ["stage/README.md"],
+            [],
+            "napp_init_v2",
+        ],
     ],
 )
-@pytest.mark.parametrize("test_project", ["napp_init_v1", "napp_init_v2"])
 def test_nativeapp_deploy_prune(
     command,
     contains,
@@ -136,7 +191,7 @@ def test_nativeapp_deploy_prune(
         os.remove(os.path.join("app", "README.md"))
 
         # deploy
-        result = runner.invoke_with_connection(command.split())
+        result = runner.invoke_with_connection(split(command))
         assert result.exit_code == 0
         assert sanitize_deploy_output(result.output) == snapshot
 
@@ -154,8 +209,16 @@ def test_nativeapp_deploy_prune(
 
 # Tests a simple flow of executing "snow app deploy [files]", verifying that only the specified files are synced to the stage
 @pytest.mark.integration
-@pytest.mark.parametrize("test_project", ["napp_init_v1", "napp_init_v2"])
+@pytest.mark.parametrize(
+    "command,test_project",
+    [
+        ["app deploy --no-validate", "napp_init_v1"],
+        ["app deploy --no-validate", "napp_init_v2"],
+        ["ws deploy --entity-id=pkg --no-validate", "napp_init_v2"],
+    ],
+)
 def test_nativeapp_deploy_files(
+    command,
     test_project,
     nativeapp_project_directory,
     runner,
@@ -170,11 +233,9 @@ def test_nativeapp_deploy_files(
         # sync only two specific files to stage
         result = runner.invoke_with_connection(
             [
-                "app",
-                "deploy",
+                *split(command),
                 "app/manifest.yml",
                 "app/setup_script.sql",
-                "--no-validate",
             ]
         )
         assert result.exit_code == 0
@@ -193,8 +254,16 @@ def test_nativeapp_deploy_files(
 
 # Tests that files inside of a symlinked directory are deployed
 @pytest.mark.integration
-@pytest.mark.parametrize("test_project", ["napp_init_v1", "napp_init_v2"])
+@pytest.mark.parametrize(
+    "command,test_project",
+    [
+        ["app deploy --no-validate", "napp_init_v1"],
+        ["app deploy --no-validate", "napp_init_v2"],
+        ["ws deploy --entity-id=pkg --no-validate", "napp_init_v2"],
+    ],
+)
 def test_nativeapp_deploy_nested_directories(
+    command,
     test_project,
     nativeapp_project_directory,
     runner,
@@ -210,7 +279,7 @@ def test_nativeapp_deploy_nested_directories(
         touch("app/nested/dir/file.txt")
 
         result = runner.invoke_with_connection(
-            ["app", "deploy", "app/nested/dir/file.txt", "--no-validate"]
+            [*split(command), "app/nested/dir/file.txt"]
         )
         assert result.exit_code == 0
         assert sanitize_deploy_output(result.output) == snapshot
@@ -227,8 +296,16 @@ def test_nativeapp_deploy_nested_directories(
 
 # Tests that deploying a directory recursively syncs all of its contents
 @pytest.mark.integration
-@pytest.mark.parametrize("test_project", ["napp_init_v1", "napp_init_v2"])
+@pytest.mark.parametrize(
+    "command,test_project",
+    [
+        ["app deploy --no-validate", "napp_init_v1"],
+        ["app deploy --no-validate", "napp_init_v2"],
+        ["ws deploy --entity-id=pkg --no-validate", "napp_init_v2"],
+    ],
+)
 def test_nativeapp_deploy_directory(
+    command,
     test_project,
     nativeapp_project_directory,
     runner,
@@ -240,15 +317,13 @@ def test_nativeapp_deploy_directory(
     with nativeapp_project_directory(test_project):
         touch("app/dir/file.txt")
         result = runner.invoke_with_connection(
-            ["app", "deploy", "app/dir", "--no-recursive", "--no-validate"]
+            [*split(command), "app/dir", "--no-recursive"]
         )
         assert_that_result_failed_with_message_containing(
             result, "Add the -r flag to deploy directories."
         )
 
-        result = runner.invoke_with_connection(
-            ["app", "deploy", "app/dir", "-r", "--no-validate"]
-        )
+        result = runner.invoke_with_connection([*split(command), "app/dir", "-r"])
         assert result.exit_code == 0
 
         package_name = f"{project_name}_pkg_{default_username}{resource_suffix}".upper()
@@ -261,56 +336,82 @@ def test_nativeapp_deploy_directory(
 
 # Tests that deploying a directory without specifying -r returns an error
 @pytest.mark.integration
-@pytest.mark.parametrize("test_project", ["napp_init_v1", "napp_init_v2"])
+@pytest.mark.parametrize(
+    "command,test_project",
+    [
+        ["app deploy --no-validate", "napp_init_v1"],
+        ["app deploy --no-validate", "napp_init_v2"],
+        ["ws deploy --entity-id=pkg --no-validate", "napp_init_v2"],
+    ],
+)
 def test_nativeapp_deploy_directory_no_recursive(
+    command,
     test_project,
     nativeapp_project_directory,
     runner,
 ):
     with nativeapp_project_directory(test_project):
         touch("app/nested/dir/file.txt")
-        result = runner.invoke_with_connection_json(
-            ["app", "deploy", "app/nested", "--no-validate"]
-        )
+        result = runner.invoke_with_connection_json([*split(command), "app/nested"])
         assert result.exit_code == 1, result.output
 
 
 # Tests that specifying an unknown path to deploy results in an error
 @pytest.mark.integration
-@pytest.mark.parametrize("test_project", ["napp_init_v1", "napp_init_v2"])
+@pytest.mark.parametrize(
+    "command,test_project",
+    [
+        ["app deploy --no-validate", "napp_init_v1"],
+        ["app deploy --no-validate", "napp_init_v2"],
+        ["ws deploy --entity-id=pkg --no-validate", "napp_init_v2"],
+    ],
+)
 def test_nativeapp_deploy_unknown_path(
+    command,
     test_project,
     nativeapp_project_directory,
     runner,
 ):
     with nativeapp_project_directory(test_project):
-        result = runner.invoke_with_connection_json(
-            ["app", "deploy", "does_not_exist", "--no-validate"]
-        )
+        result = runner.invoke_with_connection_json([*split(command), "does_not_exist"])
         assert result.exit_code == 1
         assert "The following path does not exist:" in result.output
 
 
 # Tests that specifying a path with no deploy artifact results in an error
 @pytest.mark.integration
-@pytest.mark.parametrize("test_project", ["napp_init_v1", "napp_init_v2"])
+@pytest.mark.parametrize(
+    "command,test_project",
+    [
+        ["app deploy --no-validate", "napp_init_v1"],
+        ["app deploy --no-validate", "napp_init_v2"],
+        ["ws deploy --entity-id=pkg --no-validate", "napp_init_v2"],
+    ],
+)
 def test_nativeapp_deploy_path_with_no_mapping(
+    command,
     test_project,
     nativeapp_project_directory,
     runner,
 ):
     with nativeapp_project_directory(test_project):
-        result = runner.invoke_with_connection_json(
-            ["app", "deploy", "snowflake.yml", "--no-validate"]
-        )
+        result = runner.invoke_with_connection_json([*split(command), "snowflake.yml"])
         assert result.exit_code == 1
         assert "No artifact found for" in result.output
 
 
 # Tests that specifying a path and pruning result in an error
 @pytest.mark.integration
-@pytest.mark.parametrize("test_project", ["napp_init_v1", "napp_init_v2"])
+@pytest.mark.parametrize(
+    "command,test_project",
+    [
+        ["app deploy", "napp_init_v1"],
+        ["app deploy", "napp_init_v2"],
+        ["ws deploy --entity-id=pkg", "napp_init_v2"],
+    ],
+)
 def test_nativeapp_deploy_rejects_pruning_when_path_is_specified(
+    command,
     test_project,
     nativeapp_project_directory,
     runner,
@@ -318,7 +419,7 @@ def test_nativeapp_deploy_rejects_pruning_when_path_is_specified(
     with nativeapp_project_directory(test_project):
         os.unlink("app/README.md")
         result = runner.invoke_with_connection_json(
-            ["app", "deploy", "app/README.md", "--prune"]
+            [*split(command), "app/README.md", "--prune"]
         )
 
         assert_that_result_is_usage_error(
@@ -330,9 +431,15 @@ def test_nativeapp_deploy_rejects_pruning_when_path_is_specified(
 # Tests that specifying a path with no direct mapping falls back to search for prefix matches
 @pytest.mark.integration
 @pytest.mark.parametrize(
-    "test_project", ["napp_deploy_prefix_matches_v1", "napp_deploy_prefix_matches_v2"]
+    "command,test_project",
+    [
+        ["app deploy", "napp_deploy_prefix_matches_v1"],
+        ["app deploy", "napp_deploy_prefix_matches_v2"],
+        ["ws deploy --entity-id=pkg", "napp_deploy_prefix_matches_v2"],
+    ],
 )
 def test_nativeapp_deploy_looks_for_prefix_matches(
+    command,
     test_project,
     nativeapp_project_directory,
     runner,
@@ -343,9 +450,8 @@ def test_nativeapp_deploy_looks_for_prefix_matches(
     sanitize_deploy_output,
 ):
     project_name = "myapp"
-
     with nativeapp_project_directory(test_project):
-        result = runner.invoke_with_connection(["app", "deploy", "-r", "app"])
+        result = runner.invoke_with_connection([*split(command), "-r", "app"])
         assert result.exit_code == 0
         assert sanitize_deploy_output(result.output) == snapshot
 
@@ -363,7 +469,7 @@ def test_nativeapp_deploy_looks_for_prefix_matches(
         )
 
         result = runner.invoke_with_connection(
-            ["app", "deploy", "-r", "lib/parent/child/c"]
+            [*split(command), "-r", "lib/parent/child/c"]
         )
         assert result.exit_code == 0
         stage_files = runner.invoke_with_connection_json(
@@ -380,7 +486,7 @@ def test_nativeapp_deploy_looks_for_prefix_matches(
         )
 
         result = runner.invoke_with_connection(
-            ["app", "deploy", "lib/parent/child/a.py"]
+            [*split(command), "lib/parent/child/a.py"]
         )
         assert result.exit_code == 0
         stage_files = runner.invoke_with_connection_json(
@@ -396,7 +502,7 @@ def test_nativeapp_deploy_looks_for_prefix_matches(
             stage_files.json, {"name": "stage/parent-lib/child/b.py"}
         )
 
-        result = runner.invoke_with_connection(["app", "deploy", "lib", "-r"])
+        result = runner.invoke_with_connection([*split(command), "lib", "-r"])
         assert result.exit_code == 0
         stage_files = runner.invoke_with_connection_json(
             ["stage", "list-files", f"{package_name}.{stage_name}"]
@@ -414,8 +520,16 @@ def test_nativeapp_deploy_looks_for_prefix_matches(
 
 # Tests that snow app deploy -r . deploys all changes
 @pytest.mark.integration
-@pytest.mark.parametrize("test_project", ["napp_init_v1", "napp_init_v2"])
+@pytest.mark.parametrize(
+    "command,test_project",
+    [
+        ["app deploy", "napp_init_v1"],
+        ["app deploy", "napp_init_v2"],
+        ["ws deploy --entity-id=pkg", "napp_init_v2"],
+    ],
+)
 def test_nativeapp_deploy_dot(
+    command,
     test_project,
     nativeapp_project_directory,
     runner,
@@ -427,7 +541,7 @@ def test_nativeapp_deploy_dot(
 ):
     project_name = "myapp"
     with nativeapp_project_directory(test_project):
-        result = runner.invoke_with_connection(["app", "deploy", "-r", "."])
+        result = runner.invoke_with_connection([*split(command), "-r", "."])
         assert result.exit_code == 0
         assert sanitize_deploy_output(result.output) == snapshot
 
@@ -439,3 +553,27 @@ def test_nativeapp_deploy_dot(
         assert contains_row_with(stage_files.json, {"name": "stage/manifest.yml"})
         assert contains_row_with(stage_files.json, {"name": "stage/setup_script.sql"})
         assert contains_row_with(stage_files.json, {"name": "stage/README.md"})
+
+
+@pytest.mark.integration
+@pytest.mark.parametrize(
+    "command,test_project",
+    [
+        ["app deploy", "napp_init_v1"],
+        ["app deploy", "napp_init_v2"],
+        ["ws deploy --entity-id=pkg", "napp_init_v2"],
+    ],
+)
+def test_nativeapp_deploy_validate_failing(
+    command, test_project, nativeapp_project_directory, runner
+):
+    with nativeapp_project_directory(test_project):
+        # Create invalid SQL file
+        Path("app/setup_script.sql").write_text("Lorem ipsum dolor sit amet")
+
+        # validate the app's setup script, this will fail
+        # because we include an empty file
+        result = runner.invoke_with_connection(split(command))
+        assert result.exit_code == 1, result.output
+        assert "Snowflake Native App setup script failed validation." in result.output
+        assert "syntax error" in result.output